Resolved all the bugs

+++ src/service/CancellationService.java
@@ -1,6 +1,7 @@
 package service;
 
 import exception.InvalidInputException;
+import exception.ItemNotFoundException;
 import exception.UnAuthorizedAccessException;
 import model.*;
 import repository.BookingRepository;
@@ -9,7 +10,6 @@ import repository.SeatRepository;
 import repository.WaitingListRepository;
 import utils.SessionStorage;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
@@ -26,38 +26,70 @@ public class CancellationService {
         this.passengerRepository = passengerRepository;
     }
 
-    public boolean cancelFullBooking(String pnr) throws NoSuchFieldException {
-        if (SessionStorage.getCurrentUser() == null) throw new UnAuthorizedAccessException("User must logIn to Cancel Booking");
+    public boolean cancelFullBooking(String pnr) {
+
         Booking booking = bookingRepository.getBookingById(pnr);
-        if (booking == null) throw new InvalidInputException("No booking found for this PNR");
+
         List<Passenger> passengers = passengerRepository.getPassengerByBooking(pnr);
-        if (passengers == null || passengers.isEmpty()) throw new NoSuchFieldException("No Passenger found");
-        List<Passenger> waitingListPassengers = new ArrayList<>();
+        if (passengers == null || passengers.isEmpty()) throw new ItemNotFoundException("No Passenger found");
+
         for (Passenger p : passengers) {
-            if (p.getSeatId() == -1){
-                waitingListPassengers.add(p);
-                continue;
-            } else {
-                seatRepository.updateStatus(p.getSeatId(), SeatStatus.AVAILABLE);
-            }
-            passengerRepository.updatePassengerStatusAndSeatNo(p.getPassengerId(), PassengerStatus.CANCELLED, -1);
+            cancelPassenger(p, booking.getTrainId());
         }
-        updateAndReassignWaitingList(waitingListPassengers, booking.getTrainId());
+
+        promoteWlListToCnf(booking.getTrainId());
+
         bookingRepository.updateStatus(pnr, BookingStatus.CANCELLED);
         return true;
     }
 
-    public void updateAndReassignWaitingList(List<Passenger> wlPassengers, int trainId) {
-        for (Passenger p : wlPassengers) {
-            if (p.getStatus().equals(PassengerStatus.RAC)) {
-                waitingListRepository.removeRacPassenger(p.getPassengerId(), trainId);
-            } else if (p.getStatus().equals(PassengerStatus.WL)) {
-                waitingListRepository.removeWlPassenger(p.getPassengerId(), trainId);
+    private void cancelPassenger(Passenger p, int trainId) {
+        if (p.getStatus().equals(PassengerStatus.CANCELLED)) return;
+
+        if (p.getStatus().equals(PassengerStatus.CNF)) {
+            if (p.getSeatId() != -1) {
+                seatRepository.updateStatus(p.getSeatId(), SeatStatus.AVAILABLE);
             }
         }
-        updateRac(trainId);
-        updateWl(trainId);
+        if (p.getStatus().equals(PassengerStatus.RAC)) {
+            waitingListRepository.removeRacPassenger(p.getPassengerId(), trainId);
+        }
+        if (p.getStatus().equals(PassengerStatus.WL)) {
+            waitingListRepository.removeWlPassenger(p.getPassengerId(), trainId);
+        }
+        passengerRepository.updatePassengerStatusAndSeatNo(p.getPassengerId(), PassengerStatus.CANCELLED, -1);
     }
+
+    public void promoteWlListToCnf(int trainId) {
+        List<Seat> availableSeats = seatRepository.getAvailableSeats(trainId);
+        for (Seat s : availableSeats) {
+            int racPassengerId = waitingListRepository.getFirstRacPassenger(trainId);
+            if (racPassengerId != -1) {
+                passengerRepository.updatePassengerStatusAndSeatNo(racPassengerId, PassengerStatus.CNF, s.getSeatNo());
+                seatRepository.updateStatus(s.getSeatId(), SeatStatus.BOOKED);
+                continue;
+            }
+            int wlPassengerId = waitingListRepository.getFirstWlPassenger(trainId);
+            if (wlPassengerId != -1) {
+                passengerRepository.updatePassengerStatusAndSeatNo(wlPassengerId, PassengerStatus.CNF, s.getSeatNo());
+                seatRepository.updateStatus(s.getSeatId(), SeatStatus.BOOKED);
+                continue;
+            }
+            break;
+        }
+
+        promoteWlToRac(trainId);
+     }
+
+     public void promoteWlToRac(int trainId) {
+        int availableRac = waitingListRepository.getAvailableRac(trainId);
+        for (int i=0; i<availableRac; i++) {
+            int wlPassengerId = waitingListRepository.getFirstWlPassenger(trainId);
+            if (wlPassengerId == -1) break;
+            passengerRepository.updatePassengerStatusAndSeatNo(wlPassengerId, PassengerStatus.RAC, -1);
+            waitingListRepository.addRac(trainId, wlPassengerId);
+        }
+     }
 
     public boolean partialCancellation(String pnr, Set<Integer> selectedPassengers) {
         if (SessionStorage.getCurrentUser() == null) throw new UnAuthorizedAccessException("User must logIn to Cancel Booking");
@@ -65,39 +97,26 @@ public class CancellationService {
         if (booking == null) throw new InvalidInputException("No booking found for this PNR");
         if (selectedPassengers == null || selectedPassengers.isEmpty()) { throw new InvalidInputException( "Select at least one passenger" ); }
         List<Passenger> passengers = passengerRepository.getPassengerByBooking(pnr);
-        List<Passenger> waitingListPassengers = new ArrayList<>();
         for (Passenger p : passengers) {
             if (selectedPassengers.contains(p.getPassengerId())) {
-                if (p.getStatus().equals(PassengerStatus.CNF)) {
-                    if (p.getSeatId() != -1) {
-                        seatRepository.updateStatus(p.getSeatId(), SeatStatus.AVAILABLE);
-                    }
-                    passengerRepository.updatePassengerStatusAndSeatNo(p.getPassengerId(), PassengerStatus.CANCELLED, -1);
-                } else {
-                    waitingListPassengers.add(p);
-                }
+                if (p.getStatus().equals(PassengerStatus.CANCELLED)) continue;
+                cancelPassenger(p, booking.getTrainId());
             }
         }
-        updateAndReassignWaitingList(waitingListPassengers, booking.getTrainId());
+        promoteWlListToCnf(booking.getTrainId());
+        if (checkAllPassengersAreRemoved(pnr)) {
+            bookingRepository.updateStatus(pnr, BookingStatus.CANCELLED);
+        }
         return true;
     }
 
-    private void updateRac(int trainId) {
-        List<Seat> cnfSeats = seatRepository.getAvailableSeats(trainId);
-        for (Seat curSeat : cnfSeats) {
-            int curPid = waitingListRepository.getFirstRacPassenger(trainId);
-            if (curPid == -1) break;
-            passengerRepository.updatePassengerStatusAndSeatNo(curPid, PassengerStatus.CNF, curSeat.getSeatId());
-            seatRepository.updateStatus(curSeat.getSeatId(), SeatStatus.BOOKED);
+    private boolean checkAllPassengersAreRemoved(String pnr) {
+        List<Passenger> passengers = passengerRepository.getPassengerByBooking(pnr);
+        for (Passenger p : passengers) {
+            if (!p.getStatus().equals(PassengerStatus.CANCELLED)) {
+                return false;
+            }
         }
-    }
-    private void updateWl(int trainId) {
-        int AvailableRac = waitingListRepository.getAvailableRac(trainId);
-        for (int i=0; i<AvailableRac; i++) {
-            int curPid = waitingListRepository.getFirstWlPassenger(trainId);
-            if (curPid == -1) break;
-            passengerRepository.updatePassengerStatusAndSeatNo(curPid, PassengerStatus.RAC, -1);
-            waitingListRepository.addRac(trainId, curPid);
-        }
+        return true;
     }
 }
